backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.40.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.10
redis==7.1.1
pydantic==2.12.4
pydantic-settings==2.12.0
alembic==1.18.3
python-dotenv==1.2.1

# Celery for async processing
celery==5.6.1
flower==2.0.1

# HTTP client for webhooks
httpx==0.28.1

faker==26.1.0

# Auth
python-jose[cryptography]==3.3.0
